Move dol.py, rel.py to lib
